+++ app/src/main/kotlin/com/simplemobiletools/musicplayer/models/Artist.kt
@@ -1,3 +1,3 @@
 package com.simplemobiletools.musicplayer.models
 
-data class Artist(val id: Int, val title: String, var albumCnt: Int, var trackCnt: Int, var albumArtId: Long)
+data class Artist(val id: Long, val title: String, var albumCnt: Int, var trackCnt: Int, var albumArtId: Long)
